requirements.txt
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
eventlet==0.40.2
Flask==3.1.1
Flask-SocketIO==5.5.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
mysql-connector-python==9.5.0
narwhals==2.0.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==12.0.0
plotly==6.2.0
psutil==7.0.0
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-pptx==1.0.2
python-socketio==5.13.0
pytz==2025.2
requests==2.32.4
scapy==2.6.1
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
xlsxwriter==3.2.9
